Add GetRecommendedInstanceType function
This PR adds a new `GetRecommendedInstanceType` function to help us solve the issue where t2.micro and t3.micro instances are each available in some AZs, but not available in others. This module takes in a list of instance types to pick from and returns the recommended one to use, which is one that's available in all AZs in the current region.