add `pod.Status.{Reason,Message}` to `PodNotAvailable` error
When writing tests using terratest it's useful to figure why a pod
failed to start.

For example we use heavily terratest in our e2e suite. Some tests
are flaky for many possible reasons at the moment we have to switch to using:
`WaitUntilPodAvailableE` and use the error to get the pod status.

Signed-off-by: Charly Molter <charly.molter@konghq.com>