adds custom logger to kubectl options
This commit adds the custom logging facility to kubectl options and to
the Kubectl Run with output that can be used to suppress stdout from
Kubectl run if needed.

The logging facility was introduced on #510 but it missed kubectl
options.

Fixes #358 since @blame19 recalled that the behavior was missing on the
k8s module.

Signed-off-by: Anne Macedo <retpolanne@posteo.net>